# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  kicknroar_backend_internal_dto_request.UpdateUserRequest:
    description: 'User profile update request - all fields are optional. Example:
      {"full_name":"John Doe Updated","skill_level":"advanced","profile_image_url":"https://example.com/avatar.jpg","preferred_locations":["Dhanmondi","Gulshan","Banani"]}'
    properties:
      full_name:
        example: John Doe Updated
        maxLength: 100
        minLength: 2
        type: string
      preferred_locations:
        example:
        - Dhanmondi
        - Gulshan
        - Banani
        items:
          type: string
        type: array
      profile_image_url:
        example: https://example.com/avatar.jpg
        type: string
      skill_level:
        enum:
        - beginner
        - intermediate
        - advanced
        - professional
        example: advanced
        type: string
    type: object
...
  kicknroar_backend_internal_dto_response.ErrorDetail:
    description: Detailed error information
    properties:
      code:
        example: VALIDATION_ERROR
        type: string
      details:
        example:
        - email is required
        - password must be at least 8 characters
        items:
          type: string
        type: array
      message:
        example: Invalid request data
        type: string
    type: object
  kicknroar_backend_internal_dto_response.ErrorResponse:
    description: Standard error response format
    properties:
      error:
        $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorDetail'
    type: object
  kicknroar_backend_internal_dto_response.UserResponse:
    description: User profile information
    properties:
      created_at:
        example: "2024-11-25T19:00:00Z"
        type: string
      email:
        example: john.doe@example.com
        type: string
      full_name:
        example: John Doe
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      is_verified:
        example: true
        type: boolean
      phone:
        example: "+8801712345678"
        type: string
      profile_image_url:
        example: https://example.com/avatar.jpg
        type: string
      skill_level:
        example: intermediate
        type: string
    type: object
host: localhost:8000
info:
  contact:
    email: support@kickandroar.com
    name: API Support
  description: Football match coordination platform for Dhaka, Bangladesh
  license:
    name: Proprietary
  title: Kick&Roar API
  version: "1.0"
paths:
...
  /api/v1/upload/avatar:
    post:
      consumes:
      - multipart/form-data
      description: 'Upload a profile image for the authenticated user (max 5MB, supports:
        jpg, jpeg, png, gif, webp)'
      parameters:
      - description: Image file (max 5MB)
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
      security:
      - Bearer: []
      summary: Upload user avatar
      tags:
      - upload
  /api/v1/users/{id}:
    get:
      consumes:
      - application/json
      description: Get a user's profile information by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
      security:
      - Bearer: []
      summary: Get user by ID
      tags:
      - users
  /api/v1/users/{id}/stats:
    get:
      consumes:
      - application/json
      description: Get statistics for a specific user (matches played, wins, etc.)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
      security:
      - Bearer: []
      summary: Get user statistics
      tags:
      - users
  /api/v1/users/me:
    get:
      consumes:
      - application/json
      description: Get the authenticated user's profile information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
      security:
      - Bearer: []
      summary: Get current user profile
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update the authenticated user's profile information (full_name,
        skill_level, profile_image_url, preferred_locations)
      parameters:
      - description: Profile update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/kicknroar_backend_internal_dto_request.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/kicknroar_backend_internal_dto_response.ErrorResponse'
      security:
      - Bearer: []
      summary: Update current user profile
      tags:
      - users
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"